Debug

+++ src/main/java/org/mizoguchi/misaki/service/admin/impl/UserAdminServiceImpl.java
@@ -58,28 +58,38 @@ public class UserAdminServiceImpl implements UserAdminService {
     }
 
     @Override
-    public PageResult<UserAdminResponse> searchUsers(Integer pageIndex, Integer pageSize, String sortField, String sortOrder,
-                                               SearchUserAdminRequest searchUserAdminRequest) {
+    public PageResult<UserAdminResponse> searchUsers(Integer pageIndex, Integer pageSize, String sortField,
+            String sortOrder,
+            SearchUserAdminRequest searchUserAdminRequest) {
         IPage<User> users = userMapper.selectPage(new Page<>(pageIndex, pageSize), new QueryWrapper<User>()
-                        .orderBy(sortField != null, sortOrder.equalsIgnoreCase(SqlConstant.ASC), sortField)
-                        .lambda()
-                        .like(searchUserAdminRequest.getId() != null, User::getId, searchUserAdminRequest.getId())
-                        .eq(searchUserAdminRequest.getAuthRole() != null, User::getAuthRole, searchUserAdminRequest.getAuthRole())
-                        .like(searchUserAdminRequest.getEmail() != null, User::getEmail, searchUserAdminRequest.getEmail())
-                        .like(searchUserAdminRequest.getUsername() != null, User::getUsername, searchUserAdminRequest.getUsername())
-                        .eq(searchUserAdminRequest.getGender() != null, User::getGender, searchUserAdminRequest.getGender())
-                        .eq(searchUserAdminRequest.getBirthday() != null, User::getBirthday, searchUserAdminRequest.getBirthday())
-                        .like(searchUserAdminRequest.getOccupation() != null, User::getOccupation, searchUserAdminRequest.getOccupation())
-                        .like(searchUserAdminRequest.getDetail() != null, User::getDetail, searchUserAdminRequest.getDetail())
-                        .eq(searchUserAdminRequest.getLastCheckInDate() != null, User::getLastCheckInDate, searchUserAdminRequest.getLastCheckInDate())
-                        .like(searchUserAdminRequest.getLastLoginTime() != null, User::getLastLoginTime, searchUserAdminRequest.getLastLoginTime())
-                        .eq(searchUserAdminRequest.getDeletePendingFlag() != null, User::getDeletePendingFlag, searchUserAdminRequest.getDeletePendingFlag())
-                        .eq(searchUserAdminRequest.getDeleteFlag() != null, User::getDeleteFlag, searchUserAdminRequest.getDeleteFlag())
-                        .like(searchUserAdminRequest.getCreateTime() != null, User::getCreateTime, searchUserAdminRequest.getCreateTime())
-                        .like(searchUserAdminRequest.getUpdateTime() != null, User::getUpdateTime, searchUserAdminRequest.getUpdateTime())
-        );
-        
-        PageResult<UserAdminResponse>  pageResult = new PageResult<>();
+                .orderBy(sortField != null, sortOrder.equalsIgnoreCase(SqlConstant.ASC), sortField)
+                .lambda()
+                .like(searchUserAdminRequest.getId() != null, User::getId, searchUserAdminRequest.getId())
+                .eq(searchUserAdminRequest.getAuthRole() != null, User::getAuthRole,
+                        searchUserAdminRequest.getAuthRole())
+                .like(searchUserAdminRequest.getEmail() != null, User::getEmail, searchUserAdminRequest.getEmail())
+                .like(searchUserAdminRequest.getUsername() != null, User::getUsername,
+                        searchUserAdminRequest.getUsername())
+                .eq(searchUserAdminRequest.getGender() != null, User::getGender, searchUserAdminRequest.getGender())
+                .eq(searchUserAdminRequest.getBirthday() != null, User::getBirthday,
+                        searchUserAdminRequest.getBirthday())
+                .like(searchUserAdminRequest.getOccupation() != null, User::getOccupation,
+                        searchUserAdminRequest.getOccupation())
+                .like(searchUserAdminRequest.getDetail() != null, User::getDetail, searchUserAdminRequest.getDetail())
+                .eq(searchUserAdminRequest.getLastCheckInDate() != null, User::getLastCheckInDate,
+                        searchUserAdminRequest.getLastCheckInDate())
+                .like(searchUserAdminRequest.getLastLoginTime() != null, User::getLastLoginTime,
+                        searchUserAdminRequest.getLastLoginTime())
+                .eq(searchUserAdminRequest.getDeletePendingFlag() != null, User::getDeletePendingFlag,
+                        searchUserAdminRequest.getDeletePendingFlag())
+                .eq(searchUserAdminRequest.getDeleteFlag() != null, User::getDeleteFlag,
+                        searchUserAdminRequest.getDeleteFlag())
+                .like(searchUserAdminRequest.getCreateTime() != null, User::getCreateTime,
+                        searchUserAdminRequest.getCreateTime())
+                .like(searchUserAdminRequest.getUpdateTime() != null, User::getUpdateTime,
+                        searchUserAdminRequest.getUpdateTime()));
+
+        PageResult<UserAdminResponse> pageResult = new PageResult<>();
         pageResult.setList(users.getRecords().stream()
                 .map(user -> {
                     UserAdminResponse userAdminResponse = new UserAdminResponse();
@@ -106,8 +116,10 @@ public class UserAdminServiceImpl implements UserAdminService {
 
         User user = new User();
         BeanUtils.copyProperties(updateUserAdminRequest, user);
-        String encryptPassword = passwordEncoder.encode(updateUserAdminRequest.getPassword());
-        user.setPassword(encryptPassword);
+        if (updateUserAdminRequest.getPassword() != null) {
+            String encryptPassword = passwordEncoder.encode(updateUserAdminRequest.getPassword());
+            user.setPassword(encryptPassword);
+        }
         user.setId(userId);
         int affectedRows = userMapper.updateById(user);
 
@@ -115,7 +127,8 @@ public class UserAdminServiceImpl implements UserAdminService {
             throw new OptimisticLockFailedException(FailMessageConstant.OPTIMISTIC_LOCK_FAILED);
         }
 
-        if (StringUtils.hasText(originalAvatarPath) && StringUtils.hasText(updateUserAdminRequest.getAvatarPath())) { // TODO 检查
+        if (StringUtils.hasText(originalAvatarPath) && StringUtils.hasText(updateUserAdminRequest.getAvatarPath())
+                && !originalAvatarPath.equals(updateUserAdminRequest.getAvatarPath())) {
             String fileName = new File(originalAvatarPath).getName();
             fileService.deleteFile(fileName);
         }
@@ -130,10 +143,9 @@ public class UserAdminServiceImpl implements UserAdminService {
         wishMapper.delete(new LambdaQueryWrapper<Wish>().eq(Wish::getUserId, userId));
 
         chatMapper.selectList(new LambdaQueryWrapper<Chat>()
-                .eq(Chat::getUserId, userId)
-        ).forEach(chat -> messageMapper.delete(new LambdaQueryWrapper<Message>()
-                .eq(Message::getChatId, chat.getId())
-        ));
+                .eq(Chat::getUserId, userId)).forEach(
+                        chat -> messageMapper.delete(new LambdaQueryWrapper<Message>()
+                                .eq(Message::getChatId, chat.getId())));
         chatMapper.delete(new LambdaQueryWrapper<Chat>().eq(Chat::getUserId, userId));
         int affectedRows = userMapper.deleteById(userId);
